Conflicts:
	playless-lib/src/main/java/io/github/eterverda/playless/lib/DistReplacementPicker.java

+++ playless-cli/src/main/java/io/github/eterverda/playless/cli/DumpCommand.java
@@ -1,6 +1,5 @@
 package io.github.eterverda.playless.cli;
 
-import org.apache.commons.codec.binary.Base32;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
@@ -12,14 +11,17 @@ import java.util.zip.ZipFile;
 
 import io.github.eterverda.playless.common.Dist;
 import io.github.eterverda.playless.common.Link;
+import io.github.eterverda.playless.common.Repo;
 import io.github.eterverda.playless.common.util.DistFactories;
 import io.github.eterverda.playless.core.InitialDistFactory;
 import io.github.eterverda.playless.core.Repository;
-import io.github.eterverda.playless.core.json.JsonDistDumper;
+import io.github.eterverda.playless.core.json.JsonRepoDumper;
 import io.github.eterverda.util.checksum.Checksum;
 
 public class DumpCommand implements Command {
     private static final boolean POST_PROCESS = Boolean.valueOf("true");
+
+    private static final String ICON_FILENAME = "icon.png";
 
     @Override
     public void main(Repository repo, String[] rawArgs) {
@@ -40,56 +42,64 @@ public class DumpCommand implements Command {
 
         final InitialDistFactory factory = new InitialDistFactory(aapt);
 
-        final JsonDistDumper dumper = new JsonDistDumper(System.out);
+        final JsonRepoDumper dumper = new JsonRepoDumper(System.out);
         dumper.setPrettyPrint(pretty);
 
-        final Dist[] dists = new Dist[args.size()];
+        final Repo.Editor repo = new Repo.Editor();
 
         for (int i = 0; i < args.size(); i++) {
             String arg = args.get(i);
             final File file = new File(arg);
 
             final Dist preProcess = factory.load(file);
-            final Dist postProcess = POST_PROCESS ? postProcess(file, preProcess) : preProcess;
+            final Dist postProcess = POST_PROCESS ? postProcess(preProcess) : preProcess;
             final Dist playProcess = playful ? playProcess(postProcess) : postProcess;
 
-            dists[i] = playProcess;
+            repo.dist(playProcess);
         }
 
-        dumper.writeDecorated(dists);
+        dumper.writeDecorated(repo.build());
 
         System.out.println();
     }
 
-    private Dist postProcess(File file, Dist dist) throws IOException {
+    private Dist postProcess(Dist dist) throws IOException {
         final Dist.Editor editor = dist.edit();
 
         for (Link link : dist.links) {
-            final String rel = link.rel;
-            if (rel.startsWith(Dist.LINK_REL_ICON)) {
+            final String rel = link.rel();
+            final String href = link.href();
+            if (rel.startsWith(Dist.LINK_REL_ICON) && href.startsWith("zip:file:")) {
                 editor.unlink(link);
+
+                final int indexOfExclamation = href.lastIndexOf('!');
+
+                final File file = new File(href.substring(9, indexOfExclamation));
+                final String entryName = href.substring(indexOfExclamation + 2);
 
                 final ZipFile zip = new ZipFile(file);
-                final String entryName = link.href.substring(link.href.lastIndexOf('!') + 2);
                 final ZipEntry entry = zip.getEntry(entryName);
 
-                editor.link(rel, checksum2urn(DistFactories.loadFingerprint(zip, entry)));
+                final Checksum fingerprint = DistFactories.loadFingerprint(zip, entry);
 
-            } else if (link.rel.equals(Dist.LINK_REL_APK)) {
+                editor.link(rel, makeUrl(fingerprint, ICON_FILENAME));
+
+            } else if (rel.equals(Dist.LINK_REL_DOWNLOAD)) {
                 editor.unlink(link);
-                editor.link(Dist.LINK_REL_APK, checksum2urn(dist.version.fingerprint));
+
+                final String applicationId = dist.applicationId;
+                final String version = dist.meta.containsKey(Dist.META_VERSION_NAME) ?
+                        dist.meta.get(Dist.META_VERSION_NAME) :
+                        Integer.toString(dist.version.versionCode);
+
+                final Checksum fingerprint = dist.version.fingerprint;
+                final String filename = applicationId + "-" + version + ".apk";
+
+                editor.link(rel, makeUrl(fingerprint, filename));
             }
         }
 
         return editor.build();
-    }
-
-    @NotNull
-    private String checksum2urn(Checksum checksum) {
-        final Base32 base32 = new Base32();
-        final byte[] value = checksum.getValue();
-        final String encodedValue = base32.encodeToString(value).toLowerCase();
-        return "urn:sha1:" + encodedValue;
     }
 
     private Dist playProcess(Dist dist) {
@@ -98,8 +108,8 @@ public class DumpCommand implements Command {
         editor.timestamp(Long.MIN_VALUE);
 
         for (Link link : dist.links) {
-            final String rel = link.rel;
-            if (rel.equals(Dist.LINK_REL_APK)) {
+            final String rel = link.rel();
+            if (rel.equals(Dist.LINK_REL_DOWNLOAD)) {
                 editor.unlink(link);
                 editor.link(Dist.LINK_REL_STORE, "https://play.google.com/store/apps/details?id=" + dist.applicationId);
 
@@ -108,6 +118,8 @@ public class DumpCommand implements Command {
                 // no link back
             }
         }
+
+        editor.unmeta(Dist.META_DOWNLOAD_SIZE);
 
         return editor.build();
     }
@@ -139,5 +151,11 @@ public class DumpCommand implements Command {
             }
         }
         return null;
+    }
+
+    @NotNull
+    private static String makeUrl(Checksum fingerprint, String filename) {
+        final String encodedValue = InstallCommand.base32(fingerprint);
+        return ".playless/" + encodedValue.substring(0, 2) + "/" + encodedValue.substring(2) + "/" + filename;
     }
 }
